requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
streamlit
python-dotenv
regex
jsonschema